Melhoria de UI e criação de sistema de coleta de vida e munição

+++ src/br/com/softwargames/entities/Ammo.java
@@ -3,7 +3,7 @@ package br.com.softwargames.entities;
 import java.awt.image.BufferedImage;
 
 public class Ammo extends Entity{
-    public Ammo(double x, double y, int width, int heigth, BufferedImage sprite) {
+    public Ammo(int x, int y, int width, int heigth, BufferedImage sprite) {
         super(x, y, width, heigth, sprite);
     }
 }
