+++ src/br/com/softwargames/entities/Lifepack.java
@@ -3,7 +3,7 @@ package br.com.softwargames.entities;
 import java.awt.image.BufferedImage;
 
 public class Lifepack extends Entity{
-    public Lifepack(double x, double y, int width, int heigth, BufferedImage sprite) {
+    public Lifepack(int x, int y, int width, int heigth, BufferedImage sprite) {
         super(x, y, width, heigth, sprite);
     }
 }
