+++ src/br/com/softwargames/entities/Weapon.java
@@ -3,7 +3,7 @@ package br.com.softwargames.entities;
 import java.awt.image.BufferedImage;
 
 public class Weapon extends Entity{
-    public Weapon(double x, double y, int width, int heigth, BufferedImage sprite) {
+    public Weapon(int x, int y, int width, int heigth, BufferedImage sprite) {
         super(x, y, width, heigth, sprite);
     }
 }
